% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_through_proxy.r
\name{git_through_proxy}
\alias{git_through_proxy}
\title{Connect git to github.com through proxy}
\usage{
git_through_proxy()
}
\value{
Allows git software to talk to github.com. Only needs to be run once.
}
\description{
Connect git to github.com through proxy
}
\examples{
git_through_proxy()

}
