% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_proxy_connect.r
\name{lazy_proxy_connect}
\alias{lazy_proxy_connect}
\title{Connect through proxy with Demeter credentials in environment}
\usage{
lazy_proxy_connect(username = "m123456", password = "password")
}
\arguments{
\item{username}{Your demeter username.}

\item{password}{Your demeter password.}
}
\value{
Allows internet connectivity, e.g. APIs, webscraping, installing packages from github.
}
\description{
Connect through proxy with Demeter credentials in environment
}
\examples{
lazy_proxy_connect(username="m123456", password="password")

}
