% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy_connect.r
\name{proxy_connect}
\alias{proxy_connect}
\title{Connect through proxy with Demeter credentials hidden}
\usage{
proxy_connect(show_characters = FALSE)
}
\arguments{
\item{show_characters}{logical indicating whether the characters of the password whould be visible.}
}
\value{
Allows internet connectivity, e.g. APIs, webscraping, installing packages from github.
}
\description{
Connect through proxy with Demeter credentials hidden
}
\examples{
proxy_connect()

}
